Create framework-BETA.cmd
Vastly superior to the original framework.cmd in many ways

All you need to do now is just add the script.cmd of the patch or fix. You do not have to manually put in the commands in the framework.cmd, it is completely automated like a real plugin-framework system. And it was all painfully done in batch.

You will need Windows 7 and up, Windows 10 preferred. If you still use vista or xp or anything before you are a monkey. Cyber patriot will only use windows 10 images from now on (or atleast they say).

This is still in test mode, need bug testers ;)